requirements.txt
pandas>=1.5.0
torch>=2.0.0
chronos-forecasting>=1.0.0
yfinance>=0.2.0
matplotlib>=3.5.0
numpy>=1.23.0
pandas-datareader>=0.10.0
fredapi>=0.5.0
jupyter>=1.0.0
notebook>=6.5.0
tqdm>=4.65.0
python-dateutil>=2.8.0
